eticaret/PaymentServices/requirements.txt
iyzipay
pandas
Flask
kafka-python
